Modified cover image to scale correctly and modified title text.
Cover image will "cover" the available area - note this means stretching will occur if the image is too small for the window.

Modified the heading above the cover to be the blog title, and the text over the cover image to be the description or the post title depending on the page.

Fixes #3